refactor(sidebar): extract sidebar-helpers, conversation-item, and folder-item
Split components/sidebar.tsx into focused modules:
- lib/sidebar-helpers.ts: compareConversations, mergeConversations,
  getConversationSectionLabel, buildConversationSections, escapeHtml,
  highlightMatch (pure functions, no React)
- components/conversation-item.tsx: ConversationItem + DropdownPortal
- components/folder-item.tsx: FolderItem
- sidebar.tsx retains main Sidebar component with re-exports
