+++ components/sidebar.tsx
@@ -1,23 +1,16 @@
 "use client";
 
 import { useState, useRef, useEffect, useCallback, useLayoutEffect, type MouseEvent as ReactMouseEvent } from "react";
-import { createPortal } from "react-dom";
 import { usePathname, useRouter } from "next/navigation";
 import Link from "next/link";
 import {
   Plus,
   Search,
-  MoreHorizontal,
   FolderIcon,
-  FolderOpen,
   ChevronDown,
-  ChevronRight,
-  Trash2,
-  FolderInput,
-  Pencil,
+  LoaderCircle,
   X,
-  MessageSquare,
-  LoaderCircle
+  MessageSquare
 } from "lucide-react";
 import {
   DndContext,
@@ -35,10 +28,8 @@ import {
 } from "@dnd-kit/core";
 import {
   SortableContext,
-  useSortable,
   verticalListSortingStrategy
 } from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
 
 import {
   FOLDER_DROP_PREFIX,
@@ -60,661 +51,18 @@ import { deleteConversationIfStillEmpty } from "@/lib/conversation-drafts";
 import { addGlobalWsListener } from "@/lib/ws-client";
 import type { ServerMessage } from "@/lib/ws-protocol";
 import type { Conversation, ConversationListPage, ConversationSearchResult, Folder } from "@/lib/types";
+import {
+  compareConversations,
+  mergeConversations,
+  buildConversationSections,
+  highlightMatch,
+  type SidebarConversation
+} from "@/lib/sidebar-helpers";
+import { ConversationItem } from "@/components/conversation-item";
+import { FolderItem } from "@/components/folder-item";
 import { SidebarFooterNav } from "@/components/sidebar-footer-nav";
-import { RenameModal } from "@/components/ui/rename-modal";
 
-type SidebarConversation = ConversationSearchResult;
-
-function compareConversations(left: Conversation, right: Conversation) {
-  if (left.updatedAt === right.updatedAt) {
-    return right.id.localeCompare(left.id);
-  }
-  return left.updatedAt > right.updatedAt ? -1 : 1;
-}
-
-function mergeConversations(current: Conversation[], incoming: Conversation[]) {
-  const merged = new Map(current.map((conversation) => [conversation.id, conversation]));
-  incoming.forEach((conversation) => {
-    merged.set(conversation.id, conversation);
-  });
-  return [...merged.values()].sort(compareConversations);
-}
-
-function getConversationSectionLabel(timestamp: string, now: Date) {
-  const current = new Date(now);
-  current.setHours(0, 0, 0, 0);
-
-  const updatedAt = new Date(timestamp);
-  const updatedDay = new Date(updatedAt);
-  updatedDay.setHours(0, 0, 0, 0);
-
-  if (updatedDay.getTime() === current.getTime()) {
-    return "Today";
-  }
-
-  const yesterday = new Date(current);
-  yesterday.setDate(yesterday.getDate() - 1);
-  if (updatedDay.getTime() === yesterday.getTime()) {
-    return "Yesterday";
-  }
-
-  const weekStart = new Date(current);
-  const offset = (weekStart.getDay() + 6) % 7;
-  weekStart.setDate(weekStart.getDate() - offset);
-  if (updatedDay >= weekStart) {
-    return "This Week";
-  }
-
-  const monthStart = new Date(current.getFullYear(), current.getMonth(), 1);
-  if (updatedDay >= monthStart) {
-    return "This Month";
-  }
-
-  return "Older";
-}
-
-function buildConversationSections(conversations: SidebarConversation[]) {
-  const sections = new Map<string, SidebarConversation[]>();
-  const now = new Date();
-
-  conversations.forEach((conversation) => {
-    const label = getConversationSectionLabel(conversation.updatedAt, now);
-    const list = sections.get(label) ?? [];
-    list.push(conversation);
-    sections.set(label, list);
-  });
-
-  return ["Today", "Yesterday", "This Week", "This Month", "Older"]
-    .map((label) => ({
-      label,
-      conversations: sections.get(label) ?? []
-    }))
-    .filter((section) => section.conversations.length > 0);
-}
-
-function escapeHtml(text: string): string {
-  return text
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-export function highlightMatch(text: string, query: string): string {
-  if (!query) {
-    return escapeHtml(text);
-  }
-
-  const escapedQuery = query.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-  const splitRegex = new RegExp(`(${escapedQuery})`, "gi");
-  const wholeMatchRegex = new RegExp(`^${escapedQuery}$`, "i");
-
-  return text
-    .split(splitRegex)
-    .map((segment) =>
-      wholeMatchRegex.test(segment)
-        ? `<mark class="bg-[var(--accent)]/30 text-white rounded px-0.5">${escapeHtml(segment)}</mark>`
-        : escapeHtml(segment)
-    )
-    .join("");
-}
-
-function DropdownPortal({
-  anchorRef,
-  children,
-  open,
-}: {
-  anchorRef: React.RefObject<HTMLElement | null>;
-  children: React.ReactNode;
-  open: boolean;
-}) {
-  const [coords, setCoords] = useState<{ top: number; left: number; width: number } | null>(null);
-
-  useLayoutEffect(() => {
-    if (!open || !anchorRef.current) return;
-    const rect = anchorRef.current.getBoundingClientRect();
-    setCoords({ top: rect.bottom + 4, left: rect.right - 224, width: 224 });
-  }, [open]);
-
-  useEffect(() => {
-    if (!open) return;
-    function handleScroll() {
-      if (!anchorRef.current) return;
-      const rect = anchorRef.current.getBoundingClientRect();
-      setCoords({ top: rect.bottom + 4, left: rect.right - 224, width: 224 });
-    }
-    window.addEventListener("scroll", handleScroll, true);
-    return () => window.removeEventListener("scroll", handleScroll, true);
-  }, [open]);
-
-  if (!open || !coords) return null;
-
-  return createPortal(
-    <div style={{ position: "fixed", top: coords.top, left: coords.left, width: coords.width, zIndex: 9999 }}>
-      {children}
-    </div>,
-    document.body
-  );
-}
-
-function ConversationItem({
-  conversation,
-  active,
-  onNavigate,
-  onDeleteConversation,
-  onMoveConversation,
-  allFolders,
-  dragEnabled,
-  searchQuery
-}: {
-  conversation: SidebarConversation;
-  active: boolean;
-  onNavigate?: (conversationId: string, href: string) => void | Promise<void>;
-  onDeleteConversation: (conversationId: string) => void;
-  onMoveConversation: (conversationId: string, folderId: string | null) => void;
-  allFolders: Folder[];
-  dragEnabled: boolean;
-  searchQuery: string;
-}) {
-  const router = useRouter();
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [confirmDelete, setConfirmDelete] = useState(false);
-  const [renameOpen, setRenameOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const triggerRef = useRef<HTMLButtonElement>(null);
-
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition
-  } = useSortable({ id: conversation.id });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition
-  };
-  const trimmedSearchQuery = searchQuery.trim();
-  const highlightedTitle = trimmedSearchQuery
-    ? highlightMatch(conversation.title, trimmedSearchQuery)
-    : null;
-  const highlightedMatchSnippet =
-    trimmedSearchQuery && conversation.matchSnippet
-      ? highlightMatch(conversation.matchSnippet, trimmedSearchQuery)
-      : null;
-  const titlePaddingClass = active ? "pr-8" : "pr-8 md:pr-0 md:group-hover:pr-8";
-  const actionVisibilityClass = active
-    ? "opacity-100"
-    : "opacity-100 md:opacity-0 md:group-hover:opacity-100";
-
-  function handleNavigate(event: ReactMouseEvent<HTMLAnchorElement>) {
-    if (
-      !onNavigate ||
-      event.defaultPrevented ||
-      event.button !== 0 ||
-      event.metaKey ||
-      event.ctrlKey ||
-      event.shiftKey ||
-      event.altKey
-    ) {
-      return;
-    }
-
-    event.preventDefault();
-    void onNavigate(`/chat/${conversation.id}`, conversation.id);
-  }
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    function handleClick(e: MouseEvent) {
-      if (
-        menuRef.current && menuRef.current.contains(e.target as Node)
-      ) return;
-      if (
-        triggerRef.current && triggerRef.current.contains(e.target as Node)
-      ) return;
-      setMenuOpen(false);
-    }
-    document.addEventListener("mousedown", handleClick);
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [menuOpen]);
-
-  async function handleDelete() {
-    if (!confirmDelete) {
-      setConfirmDelete(true);
-      return;
-    }
-    const response = await fetch(`/api/conversations/${conversation.id}`, { method: "DELETE" });
-    if (response.ok) {
-      onDeleteConversation(conversation.id);
-      dispatchConversationRemoved({ conversationId: conversation.id });
-    }
-    if (active) {
-      router.push("/");
-    }
-    router.refresh();
-    setMenuOpen(false);
-    setConfirmDelete(false);
-  }
-
-  async function handleMoveToFolder(folderId: string | null) {
-    const response = await fetch(`/api/conversations/${conversation.id}`, {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ folderId })
-    });
-    if (response.ok) {
-      onMoveConversation(conversation.id, folderId);
-    }
-    router.refresh();
-    setMenuOpen(false);
-  }
-
-  async function handleRenameConversation(newTitle: string) {
-    const response = await fetch(`/api/conversations/${conversation.id}`, {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ title: newTitle })
-    });
-    if (response.ok) {
-      dispatchConversationTitleUpdated({
-        conversationId: conversation.id,
-        title: newTitle
-      });
-    }
-    router.refresh();
-  }
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className="relative mb-0.5"
-      {...(dragEnabled ? attributes : {})}
-      {...(dragEnabled ? listeners : {})}
-      aria-label={`${conversation.title} conversation`}
-    >
-      <Link
-        href={`/chat/${conversation.id}`}
-        onClick={handleNavigate}
-        className={`group relative flex items-center gap-3 rounded-2xl px-3 py-2 text-sm transition-all duration-300 ${
-          active
-            ? "bg-white/[0.05] text-white font-semibold"
-            : "text-white/70 hover:bg-white/[0.03] hover:text-white/90"
-        }`}
-      >
-        {(conversation.isActive || conversation.titleGenerationStatus === "running") ? (
-          <LoaderCircle className="h-3.5 w-3.5 shrink-0 animate-spin text-[var(--accent)]" />
-        ) : (
-          <MessageSquare className={`h-4 w-4 shrink-0 transition-opacity duration-300 ${active ? "opacity-100 text-[var(--accent)]" : "opacity-60"}`} />
-        )}
-
-        <div className="relative min-w-0 flex-1 overflow-hidden">
-          {highlightedTitle ? (
-            <div
-              className={`truncate ${titlePaddingClass}`}
-              dangerouslySetInnerHTML={{ __html: highlightedTitle }}
-            />
-          ) : (
-            <div className={`truncate ${titlePaddingClass}`}>
-              {conversation.title}
-            </div>
-          )}
-
-          {highlightedMatchSnippet ? (
-            <div
-              className={`mt-0.5 truncate pr-8 text-xs ${active ? "text-white/55" : "text-white/40 group-hover:text-white/50"}`}
-              dangerouslySetInnerHTML={{ __html: highlightedMatchSnippet }}
-            />
-          ) : null}
-
-          <div
-            data-sidebar-row-actions="conversation"
-            className={`absolute right-0 top-0 bottom-0 flex items-center bg-gradient-to-l from-transparent via-transparent to-transparent pl-4 pr-1 transition-opacity duration-300 ${actionVisibilityClass}`}
-            onClick={(e) => e.preventDefault()}
-          >
-            <button
-              ref={triggerRef}
-              type="button"
-              aria-label={`Conversation actions for ${conversation.title}`}
-              title={`Conversation actions for ${conversation.title}`}
-              className="text-white/20 hover:text-white transition-colors duration-200 p-1 rounded-lg hover:bg-white/5"
-              onClick={(e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                setMenuOpen(!menuOpen);
-                setConfirmDelete(false);
-              }}
-            >
-              <MoreHorizontal className="h-4 w-4" />
-            </button>
-          </div>
-        </div>
-      </Link>
-
-      <DropdownPortal anchorRef={triggerRef} open={menuOpen}>
-        <div
-          ref={menuRef}
-          className="w-full rounded-2xl border border-white/5 bg-[#121214] p-2 shadow-2xl backdrop-blur-xl animate-fade-in relative"
-        >
-          <button
-            onClick={() => setMenuOpen(false)}
-            className="absolute top-1.5 right-1.5 p-1 text-white/20 hover:text-white/60 transition-colors"
-          >
-            <X className="h-3.5 w-3.5" />
-          </button>
-          {confirmDelete ? (
-            <div className="px-2 py-2">
-              <p className="text-xs text-white/40 mb-3 px-1">Delete conversation?</p>
-              <div className="flex gap-2">
-                <button
-                  onClick={handleDelete}
-                  className="flex-1 rounded-xl bg-red-500/10 text-red-400 text-xs py-2.5 hover:bg-red-500/20 transition-colors duration-200 font-semibold"
-                >
-                  Delete
-                </button>
-                <button
-                  onClick={() => { setConfirmDelete(false); setMenuOpen(false); }}
-                  className="flex-1 rounded-xl bg-white/5 text-white/40 text-xs py-2.5 hover:bg-white/10 transition-colors duration-200"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          ) : (
-            <>
-              {allFolders.length > 0 && (
-                <>
-                  <button
-                    onClick={() => handleMoveToFolder(null)}
-                    className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-white/40 hover:bg-white/[0.04] hover:text-white transition-colors duration-200"
-                  >
-                    <FolderInput className="h-4 w-4 opacity-50" />
-                    No folder
-                  </button>
-                  {allFolders.map((f) => (
-                    <button
-                      key={f.id}
-                      onClick={() => handleMoveToFolder(f.id)}
-                      className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-white/40 hover:bg-white/[0.04] hover:text-white transition-colors duration-200"
-                    >
-                      <FolderIcon className="h-4 w-4 opacity-50" />
-                      {f.name}
-                    </button>
-                  ))}
-                  <div className="my-1.5 border-t border-white/5" />
-                </>
-              )}
-              <button
-                onClick={() => { setRenameOpen(true); setMenuOpen(false); }}
-                className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-white/40 hover:bg-white/[0.04] hover:text-white transition-colors duration-200"
-              >
-                <Pencil className="h-4 w-4 opacity-50" />
-                Rename
-              </button>
-              <button
-                onClick={handleDelete}
-                className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-red-400/80 hover:bg-red-500/10 hover:text-red-400 transition-colors duration-200"
-              >
-                <Trash2 className="h-4 w-4 opacity-70" />
-                Delete
-              </button>
-            </>
-          )}
-        </div>
-      </DropdownPortal>
-      <RenameModal
-        open={renameOpen}
-        onOpenChange={setRenameOpen}
-        value={conversation.title}
-        onSave={handleRenameConversation}
-        title="Rename conversation"
-      />
-    </div>
-  );
-}
-
-function FolderItem({
-  folder,
-  conversations,
-  activeConversationId,
-  allFolders,
-  onNavigate,
-  onDeleteConversation,
-  onMoveConversation,
-  onCreateInFolder,
-  dragEnabled,
-  showCount,
-  searchQuery
-}: {
-  folder: Folder;
-  conversations: SidebarConversation[];
-  activeConversationId: string | null;
-  allFolders: Folder[];
-  onNavigate?: (conversationId: string, href: string) => void | Promise<void>;
-  onDeleteConversation: (conversationId: string) => void;
-  onMoveConversation: (conversationId: string, folderId: string | null) => void;
-  onCreateInFolder: (folderId: string) => void;
-  dragEnabled: boolean;
-  showCount: boolean;
-  searchQuery: string;
-}) {
-  const [collapsed, setCollapsed] = useState(() => {
-    if (activeConversationId === null) return true;
-    return !conversations.some(c => c.id === activeConversationId);
-  });
-  const [renameOpen, setRenameOpen] = useState(false);
-  const [folderMenuOpen, setFolderMenuOpen] = useState(false);
-  const [confirmDeleteFolder, setConfirmDeleteFolder] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const triggerRef = useRef<HTMLButtonElement>(null);
-  const router = useRouter();
-
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition
-  } = useSortable({ id: folder.id });
-  const {
-    setNodeRef: setFolderDropRef,
-    isOver: isOverFolderDrop
-  } = useDroppable({
-    id: `${FOLDER_DROP_PREFIX}${folder.id}`
-  });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition
-  };
-
-  useEffect(() => {
-    if (!folderMenuOpen) return;
-    function handleClick(e: MouseEvent) {
-      if (
-        menuRef.current && menuRef.current.contains(e.target as Node)
-      ) return;
-      if (
-        triggerRef.current && triggerRef.current.contains(e.target as Node)
-      ) return;
-      setFolderMenuOpen(false);
-    }
-    document.addEventListener("mousedown", handleClick);
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [folderMenuOpen]);
-
-  useEffect(() => {
-    if (activeConversationId === null) return;
-    setCollapsed(!conversations.some(c => c.id === activeConversationId));
-  }, [activeConversationId]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  async function handleRename(newName: string) {
-    await fetch(`/api/folders/${folder.id}`, {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ name: newName })
-    });
-    router.refresh();
-  }
-
-  async function handleDeleteFolder() {
-    if (!confirmDeleteFolder) {
-      setConfirmDeleteFolder(true);
-      return;
-    }
-    await fetch(`/api/folders/${folder.id}`, { method: "DELETE" });
-    router.refresh();
-    setFolderMenuOpen(false);
-    setConfirmDeleteFolder(false);
-  }
-
-  return (
-    <div ref={setNodeRef} style={style} {...(dragEnabled ? attributes : {})}>
-      <div
-        ref={dragEnabled ? setFolderDropRef : undefined}
-        className={`group flex items-center gap-3 rounded-2xl pl-1 pr-3 py-2 text-sm text-white/70 hover:bg-white/[0.03] transition-all duration-300 cursor-pointer ${
-          dragEnabled && isOverFolderDrop
-            ? "bg-white/[0.05] border border-white/10 shadow-2xl"
-            : "border border-transparent"
-        }`}
-        data-folder-drop-id={folder.id}
-        data-folder-name={folder.name}
-        aria-label={`${folder.name} folder`}
-        {...(dragEnabled ? listeners : {})}
-      >
-        {collapsed ? (
-          <FolderIcon className="h-4 w-4 opacity-60" />
-        ) : (
-          <FolderOpen className="h-4 w-4 text-[var(--accent)] opacity-60" />
-        )}
-        <span
-          className="flex-1 truncate font-medium"
-          onClick={() => setCollapsed(!collapsed)}
-        >
-          {folder.name}
-        </span>
-        {showCount ? (
-          <span className="text-[10px] font-bold text-white/10 group-hover:text-white/20 transition-colors mr-1 tabular-nums">{conversations.length}</span>
-        ) : null}
-        <div
-          data-sidebar-row-actions="folder"
-          className="flex items-center gap-1 opacity-100 transition-opacity duration-300 md:opacity-0 md:group-hover:opacity-100"
-          onClick={(e) => e.stopPropagation()}
-        >
-          <button
-            type="button"
-            aria-label={`New chat in ${folder.name}`}
-            onClick={() => onCreateInFolder(folder.id)}
-            className="p-1 text-white/20 hover:text-white transition-colors duration-200"
-            title={`New chat in ${folder.name}`}
-          >
-            <Plus className="h-3.5 w-3.5" />
-          </button>
-          <div>
-            <button
-              ref={triggerRef}
-              type="button"
-              aria-label={`Folder actions for ${folder.name}`}
-              title={`Folder actions for ${folder.name}`}
-              onClick={() => { setFolderMenuOpen(!folderMenuOpen); setConfirmDeleteFolder(false); }}
-              className="p-1 text-white/20 hover:text-white transition-colors duration-200"
-            >
-              <MoreHorizontal className="h-3.5 w-3.5" />
-            </button>
-          </div>
-        </div>
-        <button
-          type="button"
-          aria-label={collapsed ? `Expand ${folder.name}` : `Collapse ${folder.name}`}
-          onClick={() => setCollapsed(!collapsed)}
-          className="shrink-0 p-0.5 opacity-30 hover:opacity-100 transition-opacity"
-        >
-          {collapsed ? <ChevronRight className="h-3.5 w-3.5" /> : <ChevronDown className="h-3.5 w-3.5" />}
-        </button>
-      </div>
-
-      <DropdownPortal anchorRef={triggerRef} open={folderMenuOpen}>
-        <div
-          ref={menuRef}
-          className="w-48 rounded-2xl border border-white/5 bg-[#121214] p-2 shadow-2xl backdrop-blur-xl animate-fade-in relative"
-        >
-          <button
-            onClick={() => setFolderMenuOpen(false)}
-            className="absolute top-1.5 right-1.5 p-1 text-white/20 hover:text-white/60 transition-colors"
-          >
-            <X className="h-3.5 w-3.5" />
-          </button>
-          {confirmDeleteFolder ? (
-            <div className="px-2 py-2">
-              <p className="text-xs text-white/40 mb-3 px-1">Delete folder?</p>
-              <div className="flex gap-2">
-                <button
-                  onClick={handleDeleteFolder}
-                  className="flex-1 rounded-xl bg-red-500/10 text-red-400 text-xs py-2 hover:bg-red-500/20 transition-colors duration-200 font-semibold"
-                >
-                  Delete
-                </button>
-                <button
-                  onClick={() => { setConfirmDeleteFolder(false); setFolderMenuOpen(false); }}
-                  className="flex-1 rounded-xl bg-white/5 text-white/40 text-xs py-2 hover:bg-white/10 transition-colors duration-200"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          ) : (
-            <>
-              <button
-                onClick={() => { setRenameOpen(true); setFolderMenuOpen(false); }}
-                className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-white/40 hover:bg-white/[0.04] hover:text-white transition-colors duration-200"
-              >
-                <Pencil className="h-4 w-4 opacity-50" />
-                Rename
-              </button>
-              <button
-                onClick={handleDeleteFolder}
-                className="flex w-full items-center gap-3 rounded-xl px-3 py-2 text-sm text-red-400/80 hover:bg-red-500/10 hover:text-red-400 transition-colors duration-200"
-              >
-                <Trash2 className="h-4 w-4 opacity-70" />
-                Delete
-              </button>
-            </>
-          )}
-        </div>
-      </DropdownPortal>
-
-      <RenameModal
-        open={renameOpen}
-        onOpenChange={setRenameOpen}
-        value={folder.name}
-        onSave={handleRename}
-        title="Rename folder"
-        maxLength={100}
-      />
-
-      {!collapsed && conversations.length > 0 && (
-        <div className="ml-3 mt-1 flex flex-col border-l border-white/15 pl-1.5">
-          {conversations.map((conversation) => (
-            <ConversationItem
-              key={conversation.id}
-              conversation={conversation}
-              active={activeConversationId === conversation.id}
-              onNavigate={onNavigate}
-              onDeleteConversation={onDeleteConversation}
-              onMoveConversation={onMoveConversation}
-              allFolders={allFolders}
-              dragEnabled={dragEnabled}
-              searchQuery={searchQuery}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
+export { highlightMatch } from "@/lib/sidebar-helpers";
 
 export function Sidebar({
   conversationPage,
